services/ml-mnist/pyproject.toml
[project]
name = "ml-mnist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-cloud-storage>=3.1.0",
    "numpy>=2.3.0",
    "onnxruntime>=1.22.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "torchvision>=0.22.1",
]
